Refactoring dump_artifacts to split apart Nomad and Docker stuff. Run it thru Pants.
