--- etc/ci_scripts/BUILD
@@ -1,1 +1,0 @@
-python_library()
