+++ etc/ci_scripts/dump_artifacts/BUILD
@@ -1,0 +1,9 @@
+python_library(
+    name="dump_artifacts_lib",
+)
+
+pex_binary(
+    name="dump_artifacts",
+    entry_point="cli.py:main",
+    zip_safe=False,
+)
